+++ frontend/components/Domain/Recipe/RecipeCategoryTagSelector.vue
@@ -45,7 +45,7 @@
 <script>
 import RecipeCategoryTagDialog from "./RecipeCategoryTagDialog";
 import { useUserApi } from "~/composables/api";
-import { useTags, useCategories } from "~/composables/use-tags-categories";
+import { useTags, useCategories } from "~/composables/recipes";
 const MOUNTED_EVENT = "mounted";
 export default {
   components: {
