+++ frontend/composables/recipes/use-fraction.ts
