+++ frontend/composables/use-utils.ts
@@ -1,3 +1,10 @@
 export const useAsyncKey = function () {
   return String(Date.now());
 };
+
+export function uuid4() {
+  // @ts-ignore
+  return ([1e7] + -1e3 + -4e3 + -8e3 + -1e11).replace(/[018]/g, (c) =>
+    (c ^ (crypto.getRandomValues(new Uint8Array(1))[0] & (15 >> (c / 4)))).toString(16)
+  );
+}
